% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bball.R
\docType{data}
\name{bball}
\alias{bball}
\title{Example Basketball Results dataframe}
\format{A dataframe with 10 rows and 4 variables:
\describe{
  \item{team1}{Home team name}
  \item{team2}{Visiting team name}
  \item{t1}{Points scored by home team}
  \item{t2}{Points scored by visiting team}
}}
\usage{
bball
}
\description{
Sample results from Hypothetical Basketball tournament
}
\keyword{datasets}

