% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bball2.R
\docType{data}
\name{bball2}
\alias{bball2}
\title{Example Basketball Results dataframe for Colley Time Method}
\format{A dataframe with 13 rows and 4 variables:
\describe{
  \item{team1}{Home team name}
  \item{team2}{Visiting team name}
  \item{winner}{0 indicates team1 wins, 1 indicates team 2 wins}
  \item{time}{Time}
}}
\usage{
bball2
}
\description{
Time based results
}
\keyword{datasets}

