% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colley.R
\name{colley}
\alias{colley}
\title{Performs Colley Matrix Ranking Algorithm}
\usage{
colley(x)
}
\arguments{
\item{x}{A winner-loser matrix or dataframe}
}
\value{
A vector of ratings
}
\description{
Performs Colley Matrix Ranking Algorithm
}
\section{Further details}{

Each cell of the input matrix represents number of wins by rows
versus columns
}
\examples{
colley(div3_2012)
colley(ivyfootball)
}

