% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colley_spread.R
\name{colley_spread}
\alias{colley_spread}
\title{Performs Colley Matrix Ranking Algorithm adapting for spread}
\usage{
colley_spread(df, spreadval = NULL, adjval = NULL)
}
\arguments{
\item{df}{A winner-loser dataframe
There should be four or five columns. The first two are each
competitor in a competition. The third columns are the points
or goals scored by individual in the first columns. The
fourth column are the points/goals scored by the individual
in the second column. The optional fifth column is the points/goals
differential.}

\item{spreadval}{The points differential above which to adjust win values}

\item{adjval}{The constant value to adjust wins by if they are above \code{spreadval}}
}
\value{
A vector of ratings
}
\description{
Performs Colley Matrix Ranking Algorithm adapting for spread
}
\section{Further details}{

All loss and ties are considered to be worth 0 wins
}
\examples{
 colley_spread(div3_2012_spread, spreadval=3, adjval=1.5)
 colley_spread(ivyfootball, 10, 1.5)
}

