% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colley_ties.R
\name{colley_ties}
\alias{colley_ties}
\title{Performs Colley Matrix Ranking Algorithm allowing ties}
\usage{
colley_ties(df)
}
\arguments{
\item{df}{A winner-loser dataframe
There should be four columns. The first two are each
competitor in a competition. The third columns are the points
or goals scored by individual in the first columns. The
fourth column are the points/goals scored by the individual
in the second column.}
}
\value{
A vector of ratings
}
\description{
Performs Colley Matrix Ranking Algorithm allowing ties
}
\section{Further details}{

ties are considered to be half-wins.
}
\examples{
 colley_ties(epl2014_15)
}

