% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colley_weight.R
\name{colley_weight}
\alias{colley_weight}
\title{Performs Colley Matrix Ranking Algorithm adapting for time}
\usage{
colley_weight(df, method = "linear", ts = NULL)
}
\arguments{
\item{df}{A winner-loser dataframe
There should be four columns. The first two are each
competitor in a competition. The third columns should
have a 1 if the team in the first column is the winner and
a 0 if the team in the second column is the winner. The
fourth column is a time unit.}

\item{method}{The method to adjust weights of wins based on time.
Can be \code{linear}, \code{exp}, \code{log} or  \code{step}.}

\item{ts}{The time unit at which the step function goes from weighting
wins as 1 to 2.}
}
\value{
A vector of ratings
}
\description{
Performs Colley Matrix Ranking Algorithm adapting for time
}
\section{Further details}{

Method not yet implemented for ties.
}
\examples{
 colley_weight(bball2, 'linear')
 colley_weight(bball2, 'log')
 colley_weight(bball2, 'exp')
 colley_weight(bball2, 'step', ts=5)
}

