% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div3_2012.R
\docType{data}
\name{div3_2012}
\alias{div3_2012}
\title{Example College Football Matrix}
\format{A matrix with 5 rows and 5 columns:
\describe{
}}
\usage{
div3_2012
}
\description{
Sample results from 2012 Division III
}
\keyword{datasets}

