% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div3_2012_spread.R
\docType{data}
\name{div3_2012_spread}
\alias{div3_2012_spread}
\title{Example College Football Results dataframe}
\format{A dataframe with 10 rows and 5 variables:
\describe{
  \item{team1}{Home team name}
  \item{team2}{Visiting team name}
  \item{t1}{Points scored by home team}
  \item{t2}{Points scored by visiting team}
  \item{dif}{Points differential}
}}
\usage{
div3_2012_spread
}
\description{
Sample results from 2012 Div III season
}
\keyword{datasets}

