% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epl2014_15.R
\docType{data}
\name{epl2014_15}
\alias{epl2014_15}
\title{Example Soccer Results dataframe}
\format{A dataframe with 380 rows and 4 variables:
\describe{
  \item{team1}{Home team name}
  \item{team2}{Visiting team name}
  \item{t1}{Goals scored by home team}
  \item{t2}{Goals scored by visiting team}
}}
\usage{
epl2014_15
}
\description{
Sample results from 2014-15 EPL season
}
\keyword{datasets}

