% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivyfootball.R
\docType{data}
\name{ivyfootball}
\alias{ivyfootball}
\title{Example College Football Results dataframe}
\format{A dataframe with 12 rows and 4 variables:
\describe{
  \item{team1}{Home team name}
  \item{team2}{Visiting team name}
  \item{t1}{Points scored by home team}
  \item{t2}{Points scored by visiting team}
}}
\usage{
ivyfootball
}
\description{
Sample results from 2009 Ivy Football season
}
\keyword{datasets}

