% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massey.R
\name{massey}
\alias{massey}
\title{Performs Massey Matrix Ranking Algorithm}
\usage{
massey(df)
}
\arguments{
\item{df}{A winner-loser dataframe
There should be four or five columns. The first two are each
competitor in a competition. The third columns are the points
or goals scored by individual in the first columns. The
fourth column are the points/goals scored by the individual
in the second column. The optional fifth column is the points/goals
differential.}
}
\value{
A vector of ratings
}
\description{
Performs Massey Matrix Ranking Algorithm
}
\examples{
 massey(bball)
 round(massey(ivyfootball),2)
}

